Issue #192 - Add Ehcache.getAll unit tests
